src/backend/requirements.txt
Flask
Flask-Cors
python-dotenv
requests
PyPDF2 # For PDF text extraction
python-docx # For DOCX text extraction
Werkzeug # Flask dependency, ensure it's updated

Flask>=2.3
Flask-Cors>=4.0
Flask-SQLAlchemy>=3.1
Flask-Bcrypt>=1.0
Flask-JWT-Extended>=4.6
python-dotenv>=1.0
PyPDF2>=3.0
python-docx>=1.0      # import docx
PyMySQL>=1.1
requests>=2.31
docx2txt>=1.0


# pip install python-pptx
# pip install openpyxl